fix in init, added test method
